requirements.txt
flask==2.0.1
werkzeug==2.0.1
flask-cors==3.0.10
requests==2.26.0
numpy>=1.22.0
pillow>=9.0.0
rasterio>=1.2.10
shapely>=1.8.0
whitebox>=2.1.0
geopandas==0.14.3
fiona==1.9.5
pyproj>=3.0.0
pymbtiles==0.5.0